+++ src/App.tsx
@@ -4,7 +4,6 @@ import Header from './ui/components/Header/Header';
 import Toast from './ui/components/Toast/Toast';
 import LoadingSpinner from './ui/components/LoadingSpinner/LoadingSpinner';
 import { Global } from '@emotion/react';
-import { ProductElement } from './types/type';
 import { DropdownContainer, Section } from './App.styles';
 import Dropdown from './ui/components/Dropdown/Dropdown';
 import ProductList from './ui/components/ProductList/ProductList';
@@ -14,18 +13,14 @@ import {
   PRODUCT_SECTION_TITLE,
   SHOPPING_MALL_TITLE,
 } from './constants/shopInfoConfig';
-import { MAX_CART_ITEM_COUNT } from './constants/cartConfig';
 import { ERROR_MESSAGE } from './constants/errorMessage';
-import { addCart, removeCart } from './api/fetchCart';
 import { useProductListContext } from './context/ProductContext';
 import { useCartListContext } from './context/CartContext';
-import { getCartId } from './utils/getCartId';
 
 function App() {
   const {
     isLoading: isProductLoading,
     error: productError,
-    fetchData: fetchProductData,
     category,
     sortBy,
     handleSortPrice,
@@ -33,42 +28,11 @@ function App() {
   } = useProductListContext();
 
   const {
-    cartList,
     isLoading: isCartLoading,
     error: cartError,
-    fetchData: fetchCartData,
+    handleAddCart,
+    handleRemoveCart,
   } = useCartListContext();
-
-  const handleAddCart = async (product: ProductElement) => {
-    if (cartList?.length === MAX_CART_ITEM_COUNT) {
-      console.error(ERROR_MESSAGE.MAX_CART_ITEM);
-      // setIsError(true);
-      return;
-    }
-
-    try {
-      await addCart(product.id);
-      await fetchProductData();
-      await fetchCartData();
-    } catch (error) {
-      console.error(error);
-      // setIsError(true);
-    }
-  };
-
-  const handleRemoveCart = async (product: ProductElement) => {
-    try {
-      if (product) {
-        const cartId = getCartId(cartList, product.id);
-        await removeCart(cartId);
-        await fetchProductData();
-        await fetchCartData();
-      }
-    } catch (error) {
-      console.error(error);
-      // setIsError(true);
-    }
-  };
 
   return (
     <>
